Add blank `.index.js` to allow for webpack imports
Webpack loaders must resolve to a file--not a directory. If a directory is
imported (`import draft from 'file.draft';`), webpack will look for an
index.js in the directory. By adding this file to the directory on save,
we make it possible to use a webpack loader that will then find all
files within the draft file and include them in the compiled bundle.